Move `cmd/rarity` into `ethier rarity` command.
